Fix for http://code.google.com/p/gaiabb-olpc/issues/detail?id=29
Solution involved creating a new function langSelect() which creates the integer access reference map. This is used by the pre-existing findLangName() to determine which language this actually represents on disk. 

IF you don't get a valid value, you will be using English. Attackers can no longer tamper with the on disk name.