Fix for http://code.google.com/p/gaiabb-olpc/issues/detail?id=2
The resolution creates two new templates (so reloading templates will be required), and changes the authC.class.php file so that it doesn't try to output directly now. This provides a much better looking:

- Incorrect login 
- Max retries reached
- Insufficient time before next login

In addition, the logic for login.php has moved down a bit to ensure that shadows are available for the error messages.